Fix and refactor cmd/mysql.
